fixed map_sha1 calculation algorithm (thanks Zephyrix) fixed map_sha1 compatibility with 64 bit processors added a hardcoded emerald gardens map_sha1 value (admin games should work again)

+++ ghost/sha1.cpp
@@ -44,9 +44,9 @@ void CSHA1::Reset()
 	m_count[1] = 0;
 }
 
-void CSHA1::Transform(unsigned long state[5], unsigned char buffer[64])
+void CSHA1::Transform(uint32_t state[5], unsigned char buffer[64])
 {
-	unsigned long a = 0, b = 0, c = 0, d = 0, e = 0;
+	uint32_t a = 0, b = 0, c = 0, d = 0, e = 0;
 
 	SHA1_WORKSPACE_BLOCK* block;
 	static unsigned char workspace[64];
@@ -96,7 +96,7 @@ void CSHA1::Transform(unsigned long state[5], unsigned char buffer[64])
 // Use this function to hash in binary data and strings
 void CSHA1::Update(unsigned char* data, unsigned int len)
 {
-	unsigned long i = 0, j = 0;
+	uint32_t i = 0, j = 0;
 
 	j = (m_count[0] >> 3) & 63;
 
@@ -123,7 +123,7 @@ void CSHA1::Update(unsigned char* data, unsigned int len)
 
 void CSHA1::Final()
 {
-	unsigned long i = 0, j = 0;
+	uint32_t i = 0, j = 0;
 	unsigned char finalcount[8] = { 0, 0, 0, 0, 0, 0, 0, 0 };
 
 	for (i = 0; i < 8; i++)
